rm deprecated Xilinx cl_ext.h
Since we will install a conda package released XRT into a single prefix
with the actual Khronos headers, the use of #include_next in cl_ext_xilinx.h
is broken in that world.  The conda package silently does not include
the Xilinx cl_ext.h because it was already provided by the ocl-icd dep
and include_next won't find cl_ext.h from Khronos because it really is in
the include/CL directory where one would expect it to be.

Updated surrounding files for the removal